pyproject.toml
[project]
name = "presentai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-socketio>=5.5.1",
    "mss>=10.0.0",
    "pillow>=11.3.0",
    "pyqt6>=6.9.1",
]
